feat: 시설 관리 페이지 이용권별 리뷰 리스트 조회

+++ root/ticket-service/src/main/java/com/example/ticketservice/entity/Ticket.java
@@ -53,7 +53,7 @@ public class Ticket extends TimeStamped {
 
     private int purchaseCount;
 
-    private int score;
+    private float score;
 
     private boolean isAvailable;
